requirements.txt
flask
redis
pyyaml
mysql-connector-python